+++ custom_components/veolia/veolia_api/__init__.py
@@ -22,4 +22,4 @@ __all__ = [
     "VeoliaPortal",
 ]
 
-__version__ = "2.2.1"
+__version__ = "2.2.2"